feat: add Vec3-based positioning and mutex-guarded target claiming across multiple interceptors

+++ src/agents/agent_type.rs
@@ -16,7 +16,7 @@ impl Agent for AgentType {
         }
     }
 
-    fn id(&self) -> u32 {
+    fn id(&self) -> u8 {
         match self {
             AgentType::Patrol(p) => p.id(),
             AgentType::Interceptor(i) => i.id(),
