+++ src/agents/mod.rs
@@ -14,6 +14,6 @@ use crate::{agents::patrol_drone::EngineError, event::MissionEvent};
 
 pub trait Agent {
     fn act(&mut self, fused_input: &MissionEvent) -> Result<MissionEntry, EngineError>;
-    fn id(&self) -> u32; // helper so Coordinator can tag entries
+    fn id(&self) -> u8; // helper so Coordinator can tag entries
     fn last_seen(&self) -> Option<Instant> { None } // default: agents that don't track this just opt out
 }
